kafka_consumer/requirements.txt
python-dotenv
psycopg2-binary
sqlalchemy
databases[postgresql]
confluent-kafka